+++ Packages/com.unity.inputsystem/CHANGELOG.md
@@ -14,6 +14,8 @@ however, it has to be formatted properly to pass verification tests.
 #### Actions
 
 - Fixed missing keyboard bindings in `DefaultInputActions.inputactions` for navigation in UI.
+- Fixed actions ending up being disabled if switching to a control scheme that has no binding for the action (case 1187377).
+- Fixed part of composite not being bound leading to subsequent part bindings not being functional (case 1189867).
 
 ## [1.0.0-preview.1] - 2019-10-11